Created NotificationReceiver.kt broadcast receiver
